requirements.txt
torch
numpy<=1.26.4
torchvision
torchdyn
torch-fidelity
torchmetrics
einops
plotnine
pythae
hydra-core
pytorch_lightning
lightning
timm
pyrootutils
torch_tb_profiler
PyMySQL 
cryptography
POT
opencv-python
scikit-image
tensorboardX
nvitop
networkx
dppy
mplcyberpunk
gpytorch
pykeops
wandb
rich
nltk
toml
torch
tqdm>=4.60.0
pandas>=1.2.4
numpy>=1.18.5
scikit_learn>=0.24.2
omegaconf>=2.1.0
rdkit
deepspeed

albumentations
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
-e git+https://github.com/openai/CLIP.git@main#egg=clip
taming-transformers-rom1504
-e git+https://github.com/CompVis/latent-diffusion@main#egg=latent-diffusion

lmdb
diffusers
torchdiffeq
ml_collections
omegaconf
timm
ninja
blobfile
accelerate
einops
opencv-python
scikit-image
h5py
git+https://github.com/rtqichen/torchdiffeq
